Move data sources to their own sub-package
